🧑🏻‍💻feat: pdf 파일 업로드 구현

+++ src/main/java/com/fastcampus05/zillinks/core/util/service/s3upload/S3UploaderService.java
@@ -3,6 +3,8 @@ package com.fastcampus05.zillinks.core.util.service.s3upload;
 import com.fastcampus05.zillinks.core.exception.Exception400;
 import com.fastcampus05.zillinks.core.exception.Exception500;
 import com.fastcampus05.zillinks.core.util.dto.s3upload.S3UploadResponse;
+import com.fastcampus05.zillinks.core.util.model.s3upload.S3UploaderFile;
+import com.fastcampus05.zillinks.core.util.model.s3upload.S3UploaderFileRepository;
 import com.fastcampus05.zillinks.core.util.model.s3upload.S3UploaderRepository;
 import com.fastcampus05.zillinks.domain.model.intropage.IntroPage;
 import com.fastcampus05.zillinks.domain.model.user.User;
@@ -16,8 +18,7 @@ import org.springframework.web.multipart.MultipartFile;
 import java.lang.reflect.Field;
 import java.lang.reflect.InvocationTargetException;
 import java.lang.reflect.Method;
-import java.util.ArrayList;
-import java.util.List;
+import java.util.UUID;
 
 @Service
 @Transactional(readOnly = true)
@@ -28,38 +29,62 @@ public class S3UploaderService {
     public static final String DEFAULT_IMAGE = "https://taeheoki-bucket.s3.ap-northeast-2.amazonaws.com/upload/506b4c3a-53de-4cee-b571-ffa074f73ea9.jpg";
     private final S3UploaderRepository s3UploaderRepository;
     private final UserRepository userRepository;
-
-    @Transactional
-    public String upload(MultipartFile multipartFile) {
-        return s3UploaderRepository.upload(multipartFile, "");
-//        throw new RuntimeException("롤백 테스트");
-    }
+    private final S3UploaderFileRepository s3UploaderFileRepository;
 
     @Transactional
     public void delete(String url) {
-        log.info("url={}", url);
-        s3UploaderRepository.delete(url);
+        log.info("delete url={}", url);
+        if (url != null)
+            s3UploaderRepository.delete(url);
     }
 
     @Transactional
     public S3UploadResponse.PathResponse uploadImage(MultipartFile image, String name, String type, User user) {
         User userPS = userRepository.findById(user.getId())
                 .orElseThrow(() -> new Exception400("email", "등록되지 않은 유저입니다."));
-
         delete(getUrlByName(userPS.getIntroPage(), type));
 
-        String dir = "upload/" + name + "/" + type;
+        String dir = name + "/" + type;
         String uploadPath = DEFAULT_IMAGE;
         if (image != null) {
-            uploadPath = s3UploaderRepository.upload(image, dir);
+            String fileName = makeFilePath(image, dir, "jpg");
+            uploadPath = s3UploaderRepository.upload(image, fileName, "image");
         }
         return S3UploadResponse.PathResponse.builder()
                 .uploadPath(uploadPath)
                 .build();
     }
 
-    private String getUrlByName(IntroPage introPage, String type) {
+    @Transactional
+    public S3UploadResponse.PathResponse uploadFile(MultipartFile file, String name, String type, User user) {
+        User userPS = userRepository.findById(user.getId())
+                .orElseThrow(() -> new Exception400("email", "등록되지 않은 유저입니다."));
+        delete(getUrlByName(userPS.getIntroPage(), type));
 
+        String dir = name + "/" + type;
+        String uploadPath = DEFAULT_IMAGE;
+        if (file != null) {
+            String fileName = makeFilePath(file, dir, "pdf");
+            uploadPath = s3UploaderRepository.upload(file, fileName, "pdf");
+        }
+        return S3UploadResponse.PathResponse.builder()
+                .uploadPath(uploadPath)
+                .build();
+    }
+
+    private String makeFilePath(MultipartFile file, String dir, String ext) {
+        String fileName = "upload/" + dir + "/" + UUID.randomUUID() + "." + ext;
+        S3UploaderFile s3UploaderFile = S3UploaderFile.builder()
+                .originalPath(file.getOriginalFilename())
+                .encodingPath(fileName)
+                .build();
+        s3UploaderFileRepository.save(s3UploaderFile);
+        return fileName;
+    }
+
+    private String getUrlByName(IntroPage introPage, String type) {
+        if (introPage == null)
+            return null;
         Class<? extends IntroPage> introPageClass = introPage.getClass();
 
         Field[] fields = introPageClass.getDeclaredFields();
